requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
prometheus_client
flask